+++ src/frameworks/opengles/gles1_native.rs
@@ -4,7 +4,15 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 //! Passthrough for a native OpenGL ES 1.1 driver.
+//!
+//! Unlike for the GLES1-on-GL2 driver, there's almost no validation of
+//! arguments here, because we assume the driver is complete and the app uses it
+//! correctly. The exception is where we expect an extension could be used that
+//! the driver might not support (e.g. vendor-specific texture compression).
+//! In such cases, we should reject vendor-specific things unless we've made
+//! sure we can emulate them on all host platforms for touchHLE.
 
+use super::util::try_decode_pvrtc;
 use super::GLES;
 use crate::window::gles11;
 use crate::window::gles11::types::*;
@@ -288,6 +296,34 @@ impl GLES for GLES1Native {
             type_,
             pixels,
         )
+    }
+    unsafe fn CompressedTexImage2D(
+        &mut self,
+        target: GLenum,
+        level: GLint,
+        internalformat: GLenum,
+        width: GLsizei,
+        height: GLsizei,
+        border: GLint,
+        image_size: GLsizei,
+        data: *const GLvoid,
+    ) {
+        let data = unsafe { std::slice::from_raw_parts(data.cast::<u8>(), image_size as usize) };
+        // IMG_texture_compression_pvrtc (only on Imagination/Apple GPUs)
+        // TODO: It would be more efficient to use hardware decoding where
+        // available (I just don't have a suitable device to try this on)
+        if !try_decode_pvrtc(
+            self,
+            target,
+            level,
+            internalformat,
+            width,
+            height,
+            border,
+            data,
+        ) {
+            unimplemented!("CompressedTexImage2D internalformat: {:#x}", internalformat);
+        }
     }
     unsafe fn CopyTexImage2D(
         &mut self,